Changet time type to double, added spawns of players and enemies.

+++ src/Game/GameObjects/beton_wall.cpp
@@ -70,4 +70,4 @@ void BetonWall::Render() const {
 }
 
 
-void BetonWall::Update(const uint64_t delta) {}
+void BetonWall::Update(const double delta) {}
