+++ src/Game/GameObjects/eagle.cpp
@@ -18,4 +18,4 @@ void Eagle::Render() const {
 	sprites_[static_cast<size_t>(state_)]->Render(position_, size_, rotation_,layer_);
 }
 
-void Eagle::Update(const uint64_t delta) {}
+void Eagle::Update(const double delta) {}
