+++ src/Renderer/sprite_animator.h
@@ -9,12 +9,12 @@ public:
 	size_t GetCurrentFrame() const{
 		return current_frame_;
 	}
-	void Update(const uint64_t delta);
+	void Update(const double delta);
 private:
 	std::shared_ptr<Sprite> sprite_;
 	size_t current_frame_;
-	uint64_t current_frame_duration_;
-	uint64_t current_animation_time_;
+	double current_frame_duration_;
+	double current_animation_time_;
 };
 
 }
